+++ libft/libft.h
@@ -85,7 +85,7 @@ void				ft_lstdel(t_list **alst, void (*del)(void *, size_t));
 void				ft_lstadd(t_list **alst, t_list *new);
 void				ft_lstiter(t_list *lst, void (*f)(t_list *elem));
 t_list				*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem));
-char				*ft_itoa_base(int value, int base);
+char				*ft_itoa_base(long long int value, int base);
 char				*ft_strlowcase(char *str);
 size_t				ft_list_size(t_list *begin_list);
 t_list				*ft_list_last(t_list *begin_list);
